pyproject.toml
[project]
name = "Workshop-DiscordBot"
version = "0.1.0"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "discord-py[speed]~=2.5.2",
    "python-dotenv~=1.0.1",
    "tzdata; os_name == 'nt'",
]
